requirements.txt
pandas
openpyxl
jupyter
mysql-connector-python